Make vim use hybrid line numbers
Previously, vim was configured to show the corresponding line number
in front of each line.

Now, additionally enable relative line numbers. This means that,
still, in front of the current line the absolute number of this line
in the file is displayed. In front of all the other lines, the
distance from the current line is shown.

This allows to easily jump to a specific line by typing the number
displayed in front of the line before typing the up or down motion.